fixed LengthRatio and ReverseLengthRatio to behave as python2 ; reintroduced Lang_Identifier

+++ filters/LengthRatio/LengthRatio.py
@@ -55,7 +55,7 @@ class LengthRatio(AbstractFilter):
 
 			f.close()
 			if self.model_exist:
-				print("Loaded stats from the model file.")
+				print ("Loaded stats from the model file.")
 
 		if extra_args['emit scores'] == True:
 			self.num_of_scans = 1
@@ -112,7 +112,7 @@ class LengthRatio(AbstractFilter):
 		return
 
 	def decide(self, tu):
-		ratio = len(tu.src_phrase) / max(len(tu.trg_phrase), 1.0)
+		ratio = int(len(tu.src_phrase) / max(len(tu.trg_phrase), 1.0))
 
 		if self.normalize:
 			ratio = round(ratio, 3)
